requirements.txt
aerich==0.7.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncmy==0.2.9
click==8.1.7
dictdiffer==0.9.0
fastapi==0.115.2
h11==0.14.0
idna==3.10
iso8601==2.1.0
numpy==2.1.2
pandas==2.2.3
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pypika-tortoise==0.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
six==1.16.0
sniffio==1.3.1
starlette==0.39.2
tomlkit==0.13.2
tortoise-orm==0.21.7
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.1
